Cleanup Span name and filename (#8538)
Our `Span` contained following fields: name, filename, document, buffer.

- `document` was never used and never set, so I removed it,
- `name` and `filename` were both used as "the name of the buffer that
contains this span", so I consolidated them both to `filename`,
- `buffer` names sense, but I added type annotation that caused mypy to
complain about it somewhere else,
- for ad-hoc queries, we were setting filename to `<string>` or
`<empty>`, which seemed hacky. Especially, because we were later
checking if filename is `<string>` so we would not pass that to clients.
So I've removed that and now it is `None` not things that do not come
from files.