Complete SQL ast parsing for ON CONFLICT (#8542)
Fill out the mapping from dicts created by libpg_query to our pg AST
for ON CONFLICT nodes.

They are quite complicated, especially the "constraint target", because
to specify an index, you can list the index columns, with sorting
modifiers.

I need this for implementing ON CONFLICT in SQL adapter, but mostly just
so I can throw "Unsupported" exceptions :D.

Most of the changes are from:
- change of type of index_elems from `list[Expr]` to `list[IndexElem]`,
which then requires a fix at all locations where we construct the node,
- ~accidental black run on the ast_builder.py~.