Expand LS testing framework and cases (#8559)
Reopen of #8558, I've accidentally closed that.

---

Language server test previously included only a basic test
that started the server and compared output to
expected string of data.

Now, we can startup a LspRunner, give it files in the workspace,
send it init messages that bring up the server, send requests
and recieve responses.

Tests for:
- basic init,
- syntax err,
- name err,
- get definition.