Support doing non-DDL DML as part of patches (#8588)
The only thing preventing this from working was two asserts, since the
delta machinery worked with non-DDL already (it had something
registered with `astnode = qlast.Query`).

Instead of doubling down on lying about the types of things, though, I
am attempting to wrap it properly in another type, which is most of
the bulk of this PR.

See #8585, where it is tested.