ON CONFLICT excluded and WHERE over SQL adapter (#8587)
Adds support for references to `excluded` rvar
in UPDATE SET and WHERE clauses. It contains
values used in the INSERT query.

This PR also fixes the "iterator" problem of the
UPDATE, which must be executed *for each*
conflicting row, unless the UPDATE WHERE clause
returns FALSE.

It implementes this by wrapping EdgeQL `update`
into a `for` loop.

Additionally, it adds a join of iterators of
"update_value" and "insert_value" relations.
Ultimately, I'm planning to use actual ids instead
of iterators when "update value" gets the id from
"insert contents" CTE. That will require more
additional work, so I'm resorting to this hack for
now.