ON CONFLICT subject rvar over SQL adapter (#8595)
This PR makes subject rvar accessible from ON CONFLICT UPDATE and WHERE
clauses.

This allows a cleanup of hack from #8587.

Specifically, this PR:
- moves "conflict update" CTE to be *after* "else" CTE (where rows with
  conflicts are computed),
- uses "else" CTE as the source of values for subject's columns,
- enforces correlation between "excluded" and "subject" rvars by
  joining on the iterator that is generated in insert value relation.
  We cannot use real ids for this because they do not exist (yet) for
  these inserted objects.
- removes the hacky pgcompiler join of "conflict update" relation and
  update "range" and replaces it by EdgeQL comparison of ids of the
  subject and update iterator.

To make all of these things work, the *resolving* part of ON CONFLICT
needs more data about the subject, dml value and pg Environment, so it
can use pg compiler's `get_path_output` machinery. This is not the ideal
apprach, because it is supposed to be isolated within pg compiler.

This PR contains a few snippets of low code quality (because they break
boundaries between abstractions and then have to depend on internal
details of the implementation). An example of this is matching the CTE
by name. Fortuantely, it is isolated in SQL DML resolver.