Improve syntactial keyword suggestion for gel-ls (#8614)
Add syntactical suggestions to gel-ls.

We did already have a basic impl of this which relied on looking at
errors produced by a parsing of the whole file.

This PR replaces that with:
- tokenizing of the whole file,
- slicing tokens before the cursor,
- parsing without error recovery,
- inspection of the parser state and extraction of possible next tokens.

This extraction looks at available actions in current state.
An important detail is that not all of these actions are valid.
They might trigger a chain of reductions that ends in a state that
does not accept the suggested token.

This approach works great for DDL and SDL which are heavily based on
keywords.
So `START` completes with `MIGRATION` and `TRANSACTION`.
See added tests for more examples.

It does not work well for completing "identifier-like" elements.
That's because of our different keyword categories.
I've made it so that when `Ident` is in suggestions, we don't suggest
unreserved keywords.
This removes all moot suggestions for `create type _`.
But it also removes suggestion `property` from within `type X { ... }`.
Also, it does get rid of reserved keywords from locations where
reserved keywords can be identifiers i.e. in `create type default::_`.

Maybe we should simulate more of action application to see if suggested
token results in any kind of ident production.
And reject only those suggestions.

Idea: we could add "suggestion priority" to grammar rules, which would
put them to the top of the suggestion list.