Try to optimize away a per-object check in access policies (#8650)
We want postgres to not be able to decide that a type rewrite CTE is
empty, because that causes it to do optimizations that break our use
of `assert_exists` for required links.

In order to prevent postgres from determining that the CTE for a type
rewrite was empty in some cases (when it has something like `global
foo` in the filter and it is false), we were injecting an extra
`.id ?= <uuid>{}` condition to each access policy filter.
This is always false but postgres doesn't know that, so it defeats
that optimization.

It also defeats the use of indexes, which is bad. Add an analysis of
the access policy query to determine whether it is safe to omit that
check (by making sure that the object itself is always referenced).

This is somewhat dodgy, though.