Semantic completion for LSP (#8632)
Currently limited to .edgeql files and suggestions of object types and
local with bindings.

How it works:
- we start with source (e.g. `select _;`, where `_` is cursor position),
- we parse the file with recovery which hopefully produces an AST with
  bogus node at the cursor location (usually `()` is injected, so we get
  `SelectQuery(result=Tuple(elements=[]))`,
- we replace the node under the cursor with `qlast.Cursor`
  (we get `SelectQuery(result=Cursor)`),
- in qlcompiler, we raise `IdentCompletionException`, which halts
  compilation and contains the list of completion items.

We need the compilation context for its state: scopes, with bindings,
partial_path_prefix, schema are all in there. This way, we will be able
to complete names of computed pointers that have been defined in the
current query.