Start invalidating evicted cache entries on other frontends (#8638)
We just weren't ever doing that. There was a commented out
signal_sysevent, but that forced a full reread, which was too
expensive to work.

That substantially *reduces* the frequency of errors
(by about two orders of magnitude in the test in #8636),
but there is still a race condition where the other frontends
may not have processed the eviction in time. So we also need
to handle the errors and cause retries, as well as evicting
the entry.

We also do debouncing and batching of the invalidations, though,
and that makes us actually require more retries in our test script. 
This is because it (by design!) delays invalidation messages, 
and so entries are more often not invalidated in time.

I think this is unlikely to matter too much in realistic workloads.

Fixes #8636.