Pass spans from parser into AST (#7704)
In `edb/common/span.py` we are using `SpanPropagator` to propagate spans
from leaf AST nodes into parents. This is slow, since it runs for
basically every node and does a full AST pass in the worst case.

With this PR, spans are passed from the parser into CST into AST.

The problem is that when multiples AST nodes are created in grammar
mapping, the span is set only on the parent:
```python
    def reduce_STAR(self, *kids):
        self.val = qlast.Path(steps=[ # we can automatically set the proper span on Path
            qlast.Splat(depth=1),     # but not on Splat
        ])
```
... so each such grammar mapping needs explict span to be set. Which is
a lot of work and a lot of low-impact code mixed with high-impact
grammar.

But the performance improvement is 150%.

Signed-off-by: aljazerzen <aljaz@edgedb.com>