Make the patch system support making tweaks to user databases (#8694)
Currently there is the edgeql+user_ext patch mode, which makes some
change to a user database when a specific extension is installed.

Generalize it to also support an edgeql+user mode, which doesn't have
an extension.

The main goal here is to be hook-up fixup functions in there, like for
https://github.com/geldata/gel/pull/8685