Gel-ls: parse on change, but compile only on save (#8689)
Follow up for #8682

That PR made is so we don't parse or compile on every document change,
but only when document is saved. This was needed because large schemas
resulted in a lot of lag of diagnostics.

This PR does parsing on every change but recompiles only on saves.
This gives syntactical diagnostics during typing. This is feasible
because they are cheap performance wise. Compilation of EdgeQL or schema
is done only on save.

But still, there is something bad going on with schema compilation.
Name errors are quick to show (<1 sec), but successful compilations are
slow. We should investigate.