Gel-ls: completion for schema files (#8704)
Adds a naive implementation of completion for schema files to gel-ls.

It works by parsing the current file with recovery, finding the node
under the cursor and determining the module context of that node.

If module is found, we suggest all types in that module as completion
items.

We also suggest module names, regardless if we are in a module or not.

This works well when completing link types or extending types.
It does not work well to complete `ObjectRef`s such as `std::in`.
It also does not complete with type names in prelude (`str`, `int64`,
...).