Move access policy filter scope subtrees to the right part of the tree (#8712)
Because of heavy uses of anchors, we compile access policy filter
scopes on their own, before then sticking them into the FILTER clause
we create.

This was causing annoying problems involving cardinality inference of
arguments that were compiled with `prefer_subquery_args`, since having
the extra BRANCH scope in the wrong place caused them to look in the
wrong places in the scope tree.

Fix this by creating a mechanism to ask for IRAnchor's to *move* their
original scope subtree to wherever in the tree they are used, instead.

There is one other place currently where we need to do this, and it
was easy to convert it to this mechanism.

We might want to use this in more places, too. (Or, better, rip out
the scope tree.)