Fix `administer schema_repair()` (#8730)
`administer schema_repair()` had a bug where, if any repairs were
done, it set the in-memory schema to be the "reloaded_schema" used to
compare against for mismatches. This schema is equivalent to the
original, except that all of the ids were different.

This meant that any queries that were compiled using that schema would
be broken, and would try to query nonexistent tables.

This only affected the `administer` codepath; running `schema_repair`
as part of patches during version upgrades still worked.

If no uncached queries were compiled before the server was restarted,
this might not be noticed, which I suspect is what happened with
previous uses of `administer schema_repair()`.

Previously schema_repair was not really tested, since it typically
only does anything when a bug resulted in a broken schema.

I've added infrastructure for intentionally breaking the schema (by
UPDATEing the reflection schema), and a couple tests.