Migrate server code to gel-rust crates (#8761)
Major refactoring to migrate these crates to those provided by
`gel-rust`. We've had all of the code that was originally spawned here
in the other project for some time now, this is just an overdue cleanup.

As part of the work, I started locking the gel-rust crates to exact
versions, which should make backporting fixes much simpler in the future
(avoiding accidental version bumps).

The only code that hasn't landed in `gel-rust` is the prototype postgres
async driver but that will likely be up in a draft PR until I have time
to fully migrate it over.