Prohibit with_config/with_globals in the testbase gel client (#8789)
I've known they were broken for a while and finally took the time to
track down why: we override encode_state in our custom subclass of the
protocol (in cython in edb.protocol.protocol.Protocol) to ignore any
passed in state. I think the reason for this is because (unlike the
stock client) our test client supports (and our tests use a lot of)
explicit CONFIGURE SESSION and SET GLOBAL commands, and it would take
a bunch of work to make those play nicely together.

(Probably not *that* much? We'd have to decode state sent by the server
and then try overlaying the context configured state.)

Anyway, prohibit it, update some comments and some call sites.