Enable indexing and json->scalar calls to be inlined by postgres (#8804)
Functions that are marked immutable but make calls to stable functions
can't be inlined, and we have a lot of functions like that. (Because
error raising is stable in postgres but we want to allow calling some
functions that can error in indexes/constraints.)

Fix this by marking the functions as stable but making sure that their
*trampoline* wrappers are immutable, since those are what actually go
into indexes/constraints.