Allow (most) unparenthesized statements in calls, etc (#8763)
Allow (most) unparenthesized statements basically everywhere we have
comma-separated lists of expressions (or `ident := expr`):
 * Function calls
 * Set literals
 * Tuples and named tuples
 * Computed pointers in shapes
 * On the right-hand-side of basically every `:=` in SDL and DDL

Annoyingly (but unlikely to matter much in practice), it can't be
*all* unparenthesized statements, because the BY clause of GROUP
allows trailing commas (and thus FOR does as well).

To deal with this, we split ExprStmt into ExprStmtSimple and
ExprStmtAnnoying, where ExprStmtSimple contains everything but GROUP
and GROUP-containing FOR.

Fixes #8176.