Support creating indexes concurrently (#8747)
The implementation in this PR works well and is solid, but there is
plenty of room for us to argue about whether the interface is good
enough.

The core idea, though, is to have an index field called
`create_concurrently` that indicates that an index should be created
concurrently, and then there will be a separate CLI command used to
actually trigger and wait for the concurrent creation.

The crux of the complication here is that CREATE INDEX CONCURRENTLY
cannot be run from inside a transaction or a pl/pgsql script, and so it
can't be transactionally packaged with anything else, like updating our
metadata.

The approach, then, is to have three separate components:
* The actual DDL, generated by `migration create`, which will put the
index
   into our schema but not actually create any indexes.
The `create_concurrently` flag is exposed, as well as a new `active`
flag
   that indicates whether the backing index actually exists.
 * `ADMINISTER concurrent_index_create(<uuid>)`, which takes the `id` of
    an index and runs `CREATE INDEX CONCURRENTLY IF NOT EXISTS` for it,
and then sets `active := true` on the schema object (and populates
postgres
    schema comments). 
It does *not* do these things transactionally, though, so we could crash
or
shutdown after creating the index but before updating the schema. That's
fine,
though: once the command is run again, it will skip the CREATE INDEX
then update
    metadata.

Fixes #8175

------

The big questions here are:
 * Is it acceptable to have create_concurrently as a schema flag?
   I chose that because users have a very schema-driven workflow,
   and the alternative seemed to be to force them to edit the migrations
   to add it.
 * Is outsourcing the state machine processing for actually creating
   the indexes to the CLI invoking ADMINISTER commands reasonable?
   It was quite straightforward to build.


I didn't try to make use of any of the `deferred` mechanism, because
that seems like it solving a different issue.