Access policies for link tables over SQL adapter (#8849)
Closes #8844

Implements access policies for link and multi-property tables over SQL
adapter.

It does this by first compiling a CTE for table's pointer's source,
using the same function we use for compiling tables of object types.

To produce a rel var of the pointer table, it joins the pointer table
with the CTE and selects all columns.