Refactor a few grammar rules (#8858)
This PR contains a few changes to the grammar, none of the should have
any real impact.

----

The original goal of this PR:

My initial goal was to simplify grammar a lot, by merging all possible
statements into one giant non-term. That statement would then be used in
the top-level block, in `CRATE MODULE`, in `CREATE TYPE` and basically
all commands that use a sub-block. That would significantly reduce the
size of the grammar, the serialized spec (which is currently 7.5MiB) and
would also reduce the memory usage of the parser.

The obvious downside would be that following statements would be
syntactically correct:

```
RENAME TO x;
ALTER BRANCH blah {
     CREATE TYPE User;
     COMMIT;
     CREATE MIGRATION X {
         DROP EXTENDING;
     };
};
```

Because of this, we would need small compiler stage after parsing (or
directly in qlcompiler) that would reject all statements that appear in
wrong contexts.

---

This plan fell into water (if this idiom does not exist in english, I
mean fell apart), because we have statements that have overlapping
syntax. For example, syntax for `EXTENDING` depends on the context, for
example:

```
CREATE TYPE A EXTENDING B;
------------------------- this is all CreateTypeStmt

ALTER TYPE A { EXTENDING B };
------------   -----------
        |               \__ this is AlterSimpleExtending
         \__ AlterTypeStmt

# So the first stmt could also be parsed as:
CREATE TYPE A EXTENDING B;
------------- -----------
        |               \__ this is AlterSimpleExtending
         \__ CreateTypeStmt
... which causes a parser conflict.
```

There are more examples of this, and my conclusion is that this is
unfixable. The language we constructed is inherently large and cannot be
deconstructed into a few simple components such as Expression, Statement
and StatementBlock, but really does need a lot of different kinds of
expressions, statements and blocks.