Remove key_vars output of graphql rewrite output (#8884)
key_vars is used to track which variables the compilation of a graphql
query might depend on. Currently, the "rewrite" step that extracts
constants will also report the key_vars that are used in
@skip/@include directives. This is an unnecessary complication,
because we now collect those during the compilation process also.

Drop this whole bit of logic, since it is a simpler flow if all of the
key_var handling is done just once.

The "downside" of this is that we will hit CacheRedirects in cases
that previously would have been handled without an extra layer of
indirection, but we are also saving having to construct several
objects that let us skip that. I don't think that matters.

I tweaked one of the tests to make sure that it failed if I removed
the @skip/@include key_var tracking from the compiler.