Add apply_access_policies_pg_default flag to Role (#8918)
This allows an individual Role to override the per-branch
configuration of apply_access_policies_pg. The point of this is so
that we can make apply_access_policies_pg the *default*, but still
make it possible to create a role where the policies aren't applied.

This is mainly for the purpose of integration with external tools that
don't provide some sort of way to inject arbitrary config commands to
the connection estabalishment.