pg-ext: fix bug in query reparse (#8938)
The bug happens when binding the same query a second time with different
settings, please refer to the added test in this PR for details.

The reason it happened is because we used the normalized query string in
the reparse, as well as the remapped `parse_data`, then it failed in the
following `remap_parameters()`.

The fix is to just use the original query, as well as the original
`parse_data`.