Improve error on COPY with globals over SQL adapter (#7860)
This PR improves error message when COPY is used on a subject that requires globals or permissions.

We translate COPY over SQL adapter to a COPY command that is issued to the database.
When the subject of the COPY is a query that uses globals or permissions, we cannot inject them
using query parameters as we do for usual queries. 
That's because COPY command [does not support query parameters](https://www.postgresql.org/message-id/flat/12415.1550157763%40sss.pgh.pa.us#682d53c41bda2d21b7cd4fba5000793c) on PostgreSQL.
Ideally, we'd want to supply globals/parameters some other way, but there is no straightforward way to do this.